pyproject.toml
[tool.poetry]
name = "audio-tail-merge"
version = "0.1.0"
description = ""
authors = ["Frederic Tschannen <fretscha@umx.net>"]

[tool.poetry.dependencies]
python = "^3.9"
ffmpeg-python = "^0.2.0"

[tool.poetry.dev-dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
